Pull random out
Random was part of the system, making it necessary to have a typed
reference. Now, it is part of (re)storable elements and can be used
arbitrarily in any entity.
Claim generation has been lifted out of the context and claim generators
now are directly matched to fit the pair return value of the underlying
reclaimable sequence.