Fix example
Example is migrated from *wep2* to *up*, Kryo serialization is properly used (I am not smart) and extended for F2D and GDX types. Observing for properties is made public. Network issues are analyzed. Repeating DSL element is allotted per-player up to 255 slots (fix for this TBD, this is a quick solution).