Allow for globals in Ents and project capture to world
To allow for non-global parameters to entities, e.g., the frontend, shell load now allows to load arguments from
a global function. This is used to push the frontend into the constructor.

Furthermore, captures are adjusted to always capture with a null argument. The intersection is then unprojected
and used as an argument to `shoot`.